refactor(frontend): split the trigger drawers and management section
Two drawer roots and the management section become composition roots plus
co-located modules. The existing shared primitives (MasterDetailRail,
useDraftMasterDetail, DrawerFooter, RunVersionField, EventSourcePicker,
SectionRail/RailField, ActiveToggle) were already used correctly, so every
extracted module keeps importing them — no new overlapping abstraction.

  TriggerSubscriptionDrawer.tsx  2089 -> 69   + drawers/subscription/ (11)
  TriggerScheduleDrawer.tsx      1153 -> 53   + drawers/schedule/ (7)
  TriggerManagementSection.tsx   1040 -> 377  + triggerManagement/ (6)

Two genuine cross-drawer duplicates move to drawers/shared/: RequiredTitle
and normalizeJson. Three near-duplicates deliberately stay per-drawer:
MAX_DRAFTS/SHOW_LIST_RAIL (same values, different documenting comments),
applicationRevisionAdapter (a module singleton — merging would change
adapter identity), and extractBoundRevId (the schedule drawer writes the
same fallback inline; replacing it would be a rewrite, not a move).

Split against current content so #5552 (unblock editing schedule and
subscription triggers) survives: extractBoundWorkflowId/isRunVersionBound,
extractDefaultBindId, the storedReferences/versionChosen derivation and the
hidden Deployed option all land in the module that now owns them. Verified
line-by-line against the upstream diff.

Public surface unchanged: both drawers keep their default export, and
useAgentTriggers / AddTriggerDropdown are re-exported from
TriggerManagementSection.tsx.
